+++ src/Model/Resolver/ContactPageConfig.php
@@ -1,41 +1,35 @@
 <?php
+
 /**
- * ScandiPWA - Progressive Web App for Magento
- *
- * Copyright © Scandiweb, Inc. All rights reserved.
+ * @category    ScandiPWA
+ * @package     ScandiPWA_ContactGraphQl
+ * @copyright   Copyright © Scandiweb, Inc. All rights reserved.
+ * @copyright   Modifications © Selveq. All rights reserved.
+ * @license     OSL-3.0 (Open Software License ("OSL") v. 3.0)
  * See LICENSE for license details.
- *
- * @license OSL-3.0 (Open Software License ("OSL") v. 3.0)
- * @package scandipwa/base-theme
- * @link https://github.com/scandipwa/base-theme
  */
 
 namespace ScandiPWA\ContactGraphQl\Model\Resolver;
 
-use Magento\Contact\Model\Config;
+use Magento\Contact\Model\ConfigInterface;
 use Magento\Framework\GraphQl\Config\Element\Field;
 use Magento\Framework\GraphQl\Query\ResolverInterface;
 use Magento\Framework\GraphQl\Schema\Type\ResolveInfo;
 
-class ContactPageConfig implements ResolverInterface {
+class ContactPageConfig implements ResolverInterface
+{
     /**
-     * @var Config
+     * @param ConfigInterface $mailConfig
      */
-    private $mailConfig;
+    public function __construct(
+        private readonly ConfigInterface $mailConfig
+    ) {}
 
     /**
-     * ContactFormConfig constructor.
-     * @param Config $mailConfig
+     * {@inheritdoc}
      */
-    public function __construct(Config $mailConfig)
+    public function resolve(Field $field, $context, ResolveInfo $info, ?array $value = null, ?array $args = null)
     {
-        $this->mailConfig = $mailConfig;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function resolve(Field $field, $context, ResolveInfo $info, array $value = null, array $args = null) {
         return [
             'enabled' => $this->mailConfig->isEnabled()
         ];